Refactor to allow generating distribution without kandel instance deployed
